fix: Move files to correct locations
This commit moves all the files to their correct locations inside the `sv-uvm-guide` directory. This should fix the build errors.
