--- prisma/schema.prisma
@@ -1,35 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-// Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
-// Try Prisma Accelerate: https://pris.ly/cli/accelerate-init
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "sqlite"
-  url      = "file:./dev.db"
-}
-
-model User {
-  id          String   @id @default(cuid())
-  //... other user fields
-  flashcards  Flashcard[]
-}
-
-model Flashcard {
-  id            String   @id @default(cuid())
-  userId        String
-  user          User     @relation(fields: [userId], references: [id])
-  topicId       String   // Corresponds to CurriculumNode.id
-  front         String   // e.g., "What is the role of uvm_driver?"
-  back          String   // The detailed answer
-  interval      Int      @default(1) // in days
-  easinessFactor Float    @default(2.5)
-  repetitions   Int      @default(0)
-  nextReviewAt  DateTime @default(now())
-
-  @@index([userId])
-}
