+++ sv-uvm-guide/prisma/schema.prisma
@@ -6,10 +6,30 @@
 
 generator client {
   provider = "prisma-client-js"
-  output   = "../src/generated/prisma"
 }
 
 datasource db {
   provider = "postgresql"
   url      = env("DATABASE_URL")
+}
+
+model User {
+  id          String   @id @default(cuid())
+  //... other user fields
+  flashcards  Flashcard[]
+}
+
+model Flashcard {
+  id            String   @id @default(cuid())
+  userId        String
+  user          User     @relation(fields: [userId], references: [id])
+  topicId       String   // Corresponds to CurriculumNode.id
+  front         String   // e.g., "What is the role of uvm_driver?"
+  back          String   // The detailed answer
+  interval      Int      @default(1) // in days
+  easinessFactor Float    @default(2.5)
+  repetitions   Int      @default(0)
+  nextReviewAt  DateTime @default(now())
+
+  @@index([userId])
 }
